refactor: Remove unused role.service.interface.ts file
The code changes remove the unused role.service.interface.ts file. This file was no longer being used in the codebase and was causing unnecessary clutter. Removing it improves code maintainability and reduces potential confusion.
