+++ src/modules/role/application/interface/role.base.interface.ts
@@ -1,0 +1,10 @@
+export interface UserRoleToTeamData {
+  teamId: string;
+  userId: string;
+  role: string;
+  id?: string;
+}
+
+export interface IUpdateUserRoleToTeamData extends UserRoleToTeamData {
+  id: string;
+}
